Move some call predicates to internal
