+++ src/rlang/lang.h
@@ -12,15 +12,6 @@
 bool r_is_call(sexp* x, const char* name);
 bool r_is_call_any(sexp* x, const char** names, int n);
 
-bool r_is_prefixed_call(sexp* x, const char* name);
-bool r_is_prefixed_call_any(sexp* x, const char ** names, int n);
-bool r_is_maybe_prefixed_call_any(sexp* x, const char ** names, int n);
-
-bool r_is_namespaced_call(sexp* x, const char* ns, const char* name);
-bool r_is_namespaced_call_any(sexp* x, const char* ns, const char** names, int n);
-
-bool r_is_special_op_call(sexp* x);
-
 sexp* r_expr_protect(sexp* x);
